+++ src/trx/game/ui/hud.h
@@ -1,4 +1,1 @@
 #pragma once
-
-#include <trx/game/ui/hud/console.h>
-#include <trx/game/ui/hud/console_logs.h>